[doc] fixed Apache vhost for 2.4.25 (Debian 9)
The configuration example simply did not work for 2.4.25 on Debian 9
(if I remember correctly, I am the author of that example): it gave a
403 Forbidden response on the /ws URI, with nothing showing up in
Apache or Buildbot logs.

Using a separate location to dispatch ProxyPass directives is a method
that I have carried out for so long I don't even remember if I
originally got it from someone else's samples or not. A fresh look at
mod_proxy documentation indicates that the ProxyPass directives are
evaluated as "first match wins"

My best guess is that the order of precedence between Location and
ProxyPass ! directives must have changed a bit between 2.4.10 and 2.4.25,
hence the latter gives the 403 before the <Location /ws> is evaluated.

The new provided example is both simpler and working in all cases.