GerritChangeSource: Python 3 compatibility
The SSH output is a byte sequence and since
GerritChangeSourceBase.lineReceived converts the JSON bytes to str
itself, buffer bytes in GerritChangeSource for Python 3 compatibility.

Add a test to check that the normal invocation does not crash. Indeed,
without the fix it crashes in Python 3. (It does not specifically check
LocalPP since that seems an implementation detail.)