waterfall: remove unnecessary horizontal scrollbar
The margin hack was added in commit v0.9.5-126-ga0f742725 and results in
an ugly scrollbar which is especially noticable with a few builders.

Add padding to the right (use 100px to match the left side) and add a
tooltip to the header item in case the title falls off the page anyway.