fix issue with ws reconnection logic breaks oauth connection sequence.
When browsing away from the page, Firefox sends the onclose event for the websocket, which eventually triggers a reload of the page.
This races with the oauth connection sequence.
As the browser do not get a document for a while during that sequence, javascript continues to run, and eventually reload the page, cancelling the oauth redirection sequence