www: add grid_view plugin
Add a grid view plugin.

It is possible to display changes in all branches or restrict to a
specific branch. It is also possible to filter builders by clicking on
builder tags.

In the settings page, the user can specify how many changes are
displayed at max plus other more internal stuff.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>