Fix ShellArg arguments being rendered only once
ShellArg arguments were rendered only once, the first time the object
was rendered. All subsequent renderings of a ShellArg instance, i.e. all
subsequent calls to the getRenderingFor() method of the instance,
computed always the same values for its arguments.

This happened because the first time a ShellArg instance was rendered it
substituted it's attributes (even the renderable ones) with their
renderings.

Fix that and have ShellArg render its arguments anew every time it's
rendered.

Signed-off-by: Nikos Tsironis <ntsironis@arrikto.com>