support cancel build during prepare worker
In latent worker scenario, depending on the actual latent implementation, worker prepare may take a long time. 
In that case, it's desirable to cancel a build during prepare worker. 
In latent worker implementation, I can add logic to check build.stopped status, and cancel preparation. But when the prepare returns false, the code here will always start a new build to retry which defeats the stop build purpose.
The change here is to really cancel the build if the build has been marked stopped.