properties: make changes info available
In addition of the master attribute, add new sourcestamps and changes
attributes to be used when the build attribute is not set.

This will be useful when calling Properties.render() outside of a build
(in schedulers for instance).

Properties.sourcestamps and Properties.changes must of course be set for
this to work. If they are unset and build is also unset, an
AttributeError will be risen when they are accessed.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>