buildrequest: add data api compat shim
Build.getAllSourceStamps() returns a list of TempSourceStamp objects
which is not consistent with what is returned by the new data API.
Modify the TempSourceStamp class so that can return an unchanged
sourcestamp dict as returned by the data API.

Do the same for what is returned by Build.allChanges() (TempChange
objects).

Also, rework the code around these classes to put the smarts into the
classes themselves.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>