pbmanager: wait for reconfig to be done before trying to auth
When restarting the master with connected workers, as soon as the TCP
port is open, the said workers will try to reconnect. This may happen
before all workers have been registered (user+password) into the
pbmanager. When that happens, we see (non fatal) errors such as follows
appear in twistd.log:

2017-11-09 11:50:14+0100 [Broker,0,1.1.0.1] invalid login from unknown user 'bot1'
2017-11-09 11:50:14+0100 [Broker,0,1.1.0.1] Peer will receive following PB traceback:
2017-11-09 11:50:14+0100 [Broker,0,1.1.0.1] Unhandled Error
        twisted.cred.error.UnauthorizedLogin:

Fix the problem by waiting for the master to be fully configured before
trying to authenticate anyone.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>