Fix version when installing from git archive
When installing from a git archive downloaded from github, the file
master/buildbot/VERSION is not present since it is generated when
building source archive with:

    ./setup.py sdist

We also do not have access to git.

To work around the problem, use the export-subst git attribute to make
git archive write the current revison information in various files and
parse this to try and detect a version.

If the archived revision is tagged, the tag will be used as version
otherwise the commit date will be used. If the code is not issued from a
`git archive`, the version will be extracted from git describe as it was
already done before.

Signed-off-by: Robin Jarry <robin@jarry.cc>