steps/source/repo.py: Use --force-sync for sync step
Pass the --force-sync option to repo sync. This ensures that if
we change a module location in the manifest then the sync will
still succeed.

This option is not the default behaviour as it can cause data loss
in the working directory but this is not a concern for buildbot
as no changes are ever made in the checked out copy.