upgrade twisted with workaround for twisted DebugInfo __del__ issue
context: when upgrading buildbot to twisted 18.7.0, the testsuite sudenly broke, with an Error coming from a previous test.

The culprit test is:

     def test_secret(self):
        command = Interpolate("echo %(secret:fuo)s")
        self.assertFailure(self.build.render(command), defer.FirstError)
        self.flushLoggedErrors(defer.FirstError)
        self.flushLoggedErrors(KeyError)
This test is depending on whether the DebugInfo has been garbage collected. If the DebugInfo has not been collected yet, the error will appear at the time of the next test.

@_oldStyle
class DebugInfo:
    """
    Deferred debug helper.
    """
    [...]

    def __del__(self):
        """
        Print tracebacks and die.

        If the *last* (and I do mean *last*) callback leaves me in an error
        state, print a traceback (if said errback is a L{Failure}).
        """
        if self.failResult is not None:
            [...]
            # this logs the failure at an undeterministic time
            log.failure(format,
                        self.failResult,
                        debugInfo=debugInfo)
This rootcause was pretty hard to find. Doing git bisect pointed to a completly unrelated commit, and it turned out that just modifying the import sequence would change gc timing, this is why we see the problem with this version of twisted.

We use this flushLoggedErrors pattern largely in Buildbot codebase (96 times)

So this is likely that this will bite us again.

My proposal for fixing it is to following patch to make sure gc happens before flushing the errors.

    def flushLoggedErrors(self, *errorTypes):
        """
        Remove stored errors received from the log.

        C{TestCase} stores each error logged during the run of the test and
        reports them as errors during the cleanup phase (after C{tearDown}).

        @param *errorTypes: If unspecified, flush all errors. Otherwise, only
        flush errors that match the given types.

        @return: A list of failures that have been removed.
        """
+       gc.collect()
        return self._observer.flushErrors(*errorTypes)