smokes: Use older chromedriver for old Chrome
Currently we always download newest chromedriver. Unfortunately it
usually supports only a couple of latest releases of Chrome which is
problematic as they get released each month.

To work around this we parse the chromedriver release notes, extract the
chrome releases that each chromedriver release supports and use the
newest for the version of Chrome on the system.