buildrequestdistributor: Don't start builds with unsatisfied locks
We have two places to check for whether a buildrequest may be built on
specific worker for builder: Builder.canStartWithWorkerForBuilder and
Builder.canStartBuild. The first checks the builder locks that will be
acquired by build, the second just calls config.canStartBuild().

The original intention before renderable builder locks have been added
in bcf85bbe59c21bb0b06e97b6370380d7b2855d08 was that
canStartWithWorkerForBuilder checks properties that are independent of
the buildrequest and canStartBuild checks properties that are dependent
on the buildrequest. At the time nothing could be really buildrequest
specific, thus there was algorithm that if all workers would fail
canStartWithWorkerForBuilder check, we would still allow rejected
workers past that check. This allows builds with known unsatisfiable
locks to be started on busy instances of Buildbot.

Once renderable builder locks have been implemented, both
canStartWithWorkerForBuilder and canStartBuild started to depend on the
actual buildrequest but that was not noticed at the time. Thus there's
no longer a reason to keep canStartWithWorkerForBuilder and
canStartBuild separate. Combining them and getting rid of the fallback
build path both avoids the unsatisfied lock issue and simplifies the
implementation.