docs/tutorial: Better images and check for §1.1.4
The Waterfall image page was incorrect; at this point in the tutorial
the actual system displays an empty page instead of the "runtests"
waterfall line.

Since the point is to confirm not only that the master is running, but
also that the worker we started is running and has connected to the
master, we instead display the Builders page which should display the
connected worker/runner.

We also replace the top page (index) image with a screenshot from the
same run so that the Buildbot's title in the image now matches what's
in the configuration file and in the second iamge.