config: add buildRequestDistributorClass
Allow using a custom BuildRequestDistributor class instead of the
default one.

This is the first step to address bug #4592. The current implementation
of the build request distributor is overly complex and suboptimal.
However, modifying the algorithm would make the BuilderConfig.nextBuild
and BuilderConfig.nextWorker functions obsolete.

There may be people using them and we cannot break the API. The idea is
to propose multiple built-in BuildRequestDistributor implementations
(will be done in separate commits).

Signed-off-by: Robin Jarry <robin@jarry.cc>