gce: test the GCE latent worker support
This commit refactors the GCE Worker / GCE Service to split
better between a representation of the GCE API (GCE service)
and a representation of the worker implementation. This was
really meant to have a simpler way to test both.

The test implementation fakes the HTTP service. It lists the
set of HTTP requests with expected responses, and makes sure
that the actual command does follow this "script".