Fix: don't start a build when monitoring a new Hg repo
In the past, the very first poll of an Hg repository would result in an
addChange() of the head revision. We can simplify the conditionals quite a bit
to avoid it.

Aside from that minor startup glitch, changing the "state" data will trigger
that condition in the future, resulting in a duplicate build (one based on the
old state data, one based on the new).

(As a side effect of the simpler code, we've also removed the "revision + 1"
arithmetic. This worked OK in the context of a single branch, but prevents us
from adding multiple branch support in the future, since rev+1 may be on a
different branch.)