db: fix integrity error during migration to schema version 52
Fix the following error when running the migration of version 52 of the
schema:

  sqlalchemy.exc.IntegrityError: (sqlite3.IntegrityError) NOT NULL constraint failed: logs.stepid
  [SQL: INSERT INTO logs SELECT * from migration_tmp]

The problem has been seen multiple times with sqlite but it could happen
with other database engines as there was no guard against such integrity
problems before version 52 of the schema.

There is nothing much we can do about it. In case such offending rows
are found, issue a warning and delete them since they are indeed
invalid.

Fixes: #4711
Fixes: 62c30b1cae28 ("db: fix foreign key constraints for "nullable" columns")
Signed-off-by: Robin Jarry <robin@jarry.cc>