Change git ssh data path to be more unique
Currently if you have multiple builders running on the same worker
with the same workdir, the ssh keys are trying to be downloaded to the
same path, which obviously causes problems with parallel builds on the same worker.

This change tries to use the builddir parameter to make the path more unique
for each builder so they won't interfere with each other.