Improve license information in setup.py
The license field should only be used when the selected license is not
available in the Trove classifiers. From the Python documentation[1]

"The license field is a text indicating the license covering the package
where the license is not a selection from the “License” Trove
classifiers."

Also, use a more precise Trove classifier.

These changes make it easier for automated tools to detect the license of the
package.

[1] https://docs.python.org/3.7/distutils/setupscript.html