Close #4844 - Error disconnecting libvirt worker
When insubstantiating a libvirt worker the disconnection will raise an
excepcion complaining that:

  builtins.TypeError: super(type, obj): obj must be an instance or subtype
  of type

This is because super() cannot find the right method to call.

But even if we call the right disconnect methon on AbstractLatentWorker
(the parent class), the effects will not be correct, as the worker, once
stopped, will not be started again until we restart the service.

This is because the disconnect method is doing things like calling
botmaster's workerLost method.

This patch refactors the stop_instance method in LibVirtWorker to skip
calling any disconnect method (just like all the other workers) and it
also introduces a escalation mechanism when doing graceful shutdowns to
call destroy on shutdown failure.