Make dev-proxy work with the trigger step
In order for trigger step sub builds UI to work, we need to parse the buildrequest URLs.
For that we add several buildbotURLs.
The one of the proxy is needed for CORS, and the one of the original buildbot.

This commits also fixes regression with angular 1.7, and do some speedups