Fix smoke test instability
This looks like an old issue. Not sure why it is seen more now, maybe because of
performance improvement of headless browser..

Problem is that protractor is supposed to wait for browser "inactive" before each command.
It actually looks for $http requests and $timeout.
This means that if we are waiting for websocket answer, then protractor will continue

This causes race conditions which makes us miss the finish of builder list load.
Thus the first successful build id is detected as 0, and the next waitForNextSuccessfulBuild will never find the build number 1