docs: unpin sphinx, fix building and improve dashboard example
sphinx was pinned due to a sphinx bug [1][2], which is fixed since
sphinx 2.2.0. However, due to another sphinx bug, `:language: guess`
no longer works [4]. As the guessed language is xml+django, which is
not as good as html+django, which highlights CSS, I choose to fix
CSS in the example html and specify the language explicitly.

[1] https://github.com/buildbot/buildbot/pull/4852
[2] https://github.com/sphinx-doc/sphinx/issues/6474
[3] https://github.com/sphinx-doc/sphinx/pull/6624
[4] https://github.com/sphinx-doc/sphinx/issues/7139