Renderable Comparison
From discussion with @cmouse on IRC, the UX is not very good when trying to compare Interpolate objects
There is a need to create a rendered function for things that can be expressed simply with operators.

The feature was actually already there but barely documented, and working only for Property object

Playing with tests, I realized that this feature of Property was actually

clashing with ComparableMixin, which resulted in unnecesary reconfiguration if Property was used in step arguments.