Fix latent workers reject builds for prop mismatch
When a LatentWorker is taken down it does not clear its property list.
Then, when a new build tries to be scheduled on that worker it might be
rejected fora property mismatch, even though it would match because the
properties would be recreated when the latent worker is prepared for the
new build.

Note, that this is only an issue when the latent worker has been stopped
and the underlying instance, e.g. docker container, has been destroyed.
Otherwise, the worker would be correct to reject the not-matching build.

This also introduces tests for the new code, and some for related old
code.