steps/trigger: Add status counts to triggered build summaries
Its useful for the build summary to state how may of the builds in a trigger
have which statuses. Add callbacks to the triggered builds so that counts
of the results can be shown to the user in the summary.

In order to be more readable, statusToString gains a count parameter so
that plural values can be shown more correctly.

The trigger unit test is updated to contain mixed success/failure counts
which also tests the plural status code.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>