docs: Use Noto Sans, Condensed font in images
Currently images use Small Capitals feature for all text, but this in
the svg file this is represented as a separate text for the first large
capital letter and a separate text for the remaining small capitals.
This makes the text very hard to work with natively within Inkscape.

Unfortunately the current "PT Sans Narrow" font does not support the
Small Capitals feature, so we switch to a similar font that does support
it.