latent: Allow to stop build while preparing latent worker
When a build is stopped while it is preparing (in case of latent workers
that means they are still substantiating) the worker needs to be
insubstantiated manually, otherwise the build does never stop.
This also requires the build's start up logic to be reworked, so that
the cancellation of the substantiation of the latent worker is not
reported as an error but as a simple cancellation.