process: Don't send builder update messages if the config didn't changed
Currently we always do a database update and send a mq message whenever
a builder is reconfigured. In most reconfigurations most builders will
stay the same. However, in Buildbot instances with many builders a large
part of reconfiguration latency comes due to these database operations.
So it's worth to optimize for the common case of no-change builder
reconfiguration.