GitHub hook: allow making certain GitHub API requests without API token
The change https://github.com/buildbot/buildbot/pull/5508 was overly
aggressive because it expected that the GitHub APIs for getting a commit
and the changed files in a pull request needed a token to function. This
is not the case for public repositories, and this change optimistically
makes the API request. The status code checking added in that previous
change handles the previous error that the change was meant to address.