latent: Don't regenerate random password when worker is substantiated
During substantiation worker is already configured to authenticate using
a specific password. If we reconfigure the worker at that time, the
password expected by the master may change in cases when random password
is reconfigured. To avoid this we check if the worker is substantiated
or in any substantiation state and use the existing password if that's
the case