git: Don't reset last active branch to wrong commit
When we change the checked out code using `git reset --hard <rev>` we
also change the revision that the currently active branch points to.
This leads to that branch pointing to a wrong commit as we're actually
checking out a different branch (we switch to it just a little bit
later). Instead of modifying the current branch we should just modify
the working copy, which can be done by `git checkout -f <rev>`.