Fix GitStepMixin._getSshDataPath() to use the builder's workerbuilddir
It was using the builder's name unmodified, which may be unsafe to use
for paths, so we make it use the builder's workerbuilddir, which is
already made safe.

This also changes the master/buildbot/test/unit/steps/test_source_git.py,
as it expected the builder name, to expect the workerbuilddir instead.

Fixes #5949