gitpoller: Exit polling early during reconfig
On larger repositories or on masters with heavy IO load gitpoller may
take a while to execute a full polling iteration. Any reconfiguration
during that time would wait for polling iteration to finish. This may
result in unacceptable delays. To solve this we add a couple of checks
in strategic places where gitpoller may exit early.