process: Fix crash when worker with a running build is disconnected
A worker disconnection will detach the worker from workerforbuilder, so
its worker member will become None. If any build was running during the
disconnection, it will no longer be able to access the worker
information. This will result in a crash as it may be used as a source
for build properties which is updated just before the end of the build.