docker: Switch to Debian and build all master packages
Previously we were installing buildbot-www from whatever is available on
Pypi. This leads to wrong versions being taken for docker containers
built from tags. E.g. if 2.10.5 is built, one expects buildbot-www to
correspond to 2.10.5 code, not what's latest on Pypi. Fortunately we can
just reuse the existing release build code to build www packages for the
container and solve the problem this way.

Current Dockerfile uses Alpine, which is great for reducing the
container footprint, but for master this is less important because by
design there will be few instances running. Additionally, most of the
users of Buildbot (myself included) are much more likely to have more
experience with Debian. Given that I need to make significant changes to
the Dockerfile anyway, I'm switching it to Debian.