process: Fix WorkerLock's maxCountForWorker when used with buildsteps
There has been an issue where the Worker object, rather than just the
worker's name, has been passed to 'RealWorkerLock.getLockForWorker'
when the lock was acquired for a single build step. As a result, the
erroneous 'workername' value was not found in the 'maxCountForWorker'
dictionary, and so we fell back to the 'maxCount' value.

We fix this by passing the workername rather than the Worker object
from 'BuildStep.startStep' into 'getLockForWorker'.