www: Don't show outdated browser warning for unindentified browsers
By showing outdated browser warning we are mostly interested in users
who are using a browser that is known to be not supported. If a browser
is not identifiable it most likely means the user is on a custom setup,
knows what he's doing and most likely has a browser that supports the
required web standards.