DOCS: Simplify 'Launching worker as Windows service'
This section used to repeat what's already described in much more and
better detail in installation.rst and worker.rst.  That's all dropped,
making this section a more true "Next Steps" section.

All that remains is the additional steps that turns an already created
buildbot worker into a Windows service.