Reconsider nextWorker for each build request
When the ability for nextWorker to make decisions based on the build
request was added, the processing of workers was not updated to ensure
it would consider each request separately. This caused the immediate
issue that popNextBuild would only consider the first build request if
nextWorker was unable to place it. Additionally, once a worker had been
asked to consider a given build request, this decision persisted to
future unrelated build requests if the build was prevented from starting
by Builder.canStartBuild.

The fix for both of these issues is always considering every available
worker for each build request as they are chosen by nextBuild.