cmake: Fix an error in passing CMake options and definitions
The cmake usage documentation (see below) dictates that the options
shall precede the <path-to-source> (and also <path-to-build>).

I don't have an isolated, standalone unit test for this, but (a) I've
personally experienced problems when forgetting this rule, and (b) the
builds on my buildbot were all ignoring the options specified in the
master.cfg file until I made this correction to cmake.py.

cmake --help
Usage

  cmake [options] <path-to-source>
  cmake [options] <path-to-existing-build>
  cmake [options] -S <path-to-source> -B <path-to-build>