test: Move contents of expect.py to steps.py
The current Expect* classes are very specific to steps testing and as a
result they should not be in a generic expect.py module.

Originally the idea was that we want to mock things that are used not
only steps, such as running commands on master. So we would have most
expectation classes in expect.py which would be used from steps and
other tests.

However, we would need to split expect.py module anyway as the test API
expands. Therefore, the current split is to put all exposed step-related
API into steps.py and have additional modules for other tests stuff.