util: Allow full URLs for HTTPClientService methods
HTTPClientService currently only supports URLs that are relative to base
URL. This is a serious limitation in case of REST APIs that return paths
to other resources in the API responses. If the paths are full URLs then
it would be simplest if we could just pass the URLs to HTTPClientService
unmodified.