worker: Change certain commands workdirs from builder to worker basedir
We don't want to depend on builder basedirs in Msgpack protocol. This is
currently only used in commands "rm", "find" and "cp". However, these
commands get all paths as absolute paths, so it does not matter what
directory they are run in. Thus we can replace builder base directory
with any other directory. We choose worker basedir because it's readily
available.