git: Do not throw on intial _fetch() failure if retry or clobber requested.
In `Git._fetchOrFallback`, the call to `self._fetch(None)` will throw a
`BuildStepFailed` if the fetch command fails. The fallback code in
`_fetchOrFallback` will not be executed and the fetch will not be
retried nor will `clobber()` be called.

This change modifies `_fetch` to accept an `abandonOnFailure` flag to
control whether it should throw.

Test Plan

Run the unit tests added in the previous commit.