Rest: Filter buildsets by sourcestamp attributes
Allow the buildsets REST endpoint to filter by sourcestamp attributes.
We use the 'contains' relation to filter by a set of sourcestamp attributes:
buildsets?sourcestamps__contains={"ssid":26}

The list of buildsets is then only listing buildsets that have a
sourcestamp with that ssid. If the sourcestamps__contains filter is used
multiple times it follows the "or" semantics like the other filters. If
a single filter lists multiple attributes these follow and semantics,
within the filter.

sourcestamps are encoded as JSON in the REST query.

The filters are implemented on SQL-level. Necessary joins are only
performed if this filter is actually used.