Add dummy workflow_run github event handler
When you add a buildbot webhook to github, github will send your
server a workflow_run event as a "ping" to sanity check that
everything is setup correctly.  Buildbot doesn't support the
workflow_run event and will return an error to github:

	400 - Unknown event: workflow_run

This commit adds in a dummy event handler for workflow_run that
returns "200 - no change found", which in turn makes github
happy.

Signed-off-by: Tony Hutter <hutter2@llnl.gov>