www/react: Add react-tabs dependency
We're still using bootstrap 3.x with react-bootstrap 1.6.x which is
compatible only with bootstrap 4.x. This is needed for tooltip widgets
to work properly.

As a result of this incompatibility the builtin react-bootstrap tab
widgets don't work. For now tabs will be implemented with react-tabs and
once bootstrap is migrated to 4.x tabs implementation will be reverted
back to components from react-bootstrap.