www/react: Replace icons from font-awesome with icons from react-icons
Font Awesome 6 includes various icons styles, some of them are included
in the free plan, others in the paid pro plan. The style that is
currently used, "regular", is in the pro plan, which means there is no
upgrade path for Font Awesome. Accordingly, a free alternative that
supports the same icon set is chosen.