www/react: Replace webpack and create-react-app with vite
create-react-app is no longer maintained. This means the webpack
configuration included from create-react-app via yarn eject is pure
technical debt as most of the features are not really needed by
BuildBot. Migrating to another build framework fixes the problem and, in
the case of vite, significantly reduces the amount of configuration that
is needed.