www/react: Change license of react-data-module to from GPL-2.0 to MPL-2.0
I'm the only contributor of this code which means I can relicense the
code under whatever license I deem necessary. The intention of
react-data-module is to be able to use it in client applications to more
easily access BuildBot Data API. Consequently, it would be difficult to
use this module if it required open-sourcing any code that touches it. A
license such as MPL-2.0 is better because it only requires publishing
changes at the source file level.