test: Make integration tests more consistent
Currently integration tests have two ways to be implemented: by deriving
RunMasterBase or deriving RunFakeMasterTestCase which work a bit
differently. However, the way the actual tests are written is also
different, which makes it harder to understand the tests as on the
surface all integration tests work similarly. This commit makes the
tests more consistent between themselves.