changes/github: Support user settable project in GitHubPullrequestPoller
When using GitHubPullrequestPoller the project is hardcoded as the
full name that is returned by the github API.

When trying to use this with existing change sources that have
specific projects, this is less than useful.

Support specifying the project as an argument when initializing
GitHubPullrequestPoller. If left unset the previous behaviour is
maintained.