process/buildrequestdistributor: sort builders based on request priority
When choosing the next builder, use the highest priority of build
requests for that builder as the priority for sorting.

Use -priority in the key as "async_sort" sorts in ascending order.

Note that this does not change default behaviour since all build
requests are created by default with priority 0 and python's sort (which
is what async_sort() eventually calls) is stable.

Update the unit tests so that builds are sorted by priority. Then add a
new test in which the priorities are the same and ensure that the
secondary key (oldestRequestTime) is used as the tie-break.