www/react: Fix queries depending on useDataApiSingleElementQuery()
The current implementation returns dummy DataCollection for the first
call (while data has not yet arrived) of useDataApiSingleElementQuery()
and similar users of useDataApiDynamicQuery(). Once the response data
arrives, the data collection stored in useDataApiQuery() is not renewed
even though it does not react to any data updates.

The fix is to treat collections with undefined accessor as invalid data
collections.