www: Fix support for very large logs in React log viewer
The current implementation creates a HTML element of height line_height
* line_count. This is problematic, because browsers do not properly
support scrolling for elements above certain size. This makes it
impossible to show logs larger than million of lines or so.

To work around this, scrolling is done in a smaller HTML element. The
scroll offset and the rows are positioned in such a way that it appears
scrolling is done on a much larger HTML element.