process/buildrequestdistributor: sort unclaimed br cache by requestid
When choosing the next buildrequest to build, we first look at their
priority, then break ties by choosing the oldest ("submitted_at" field).

However, if multiple buildrequests were created at the same time, they
will all have the same "submitted_at". This makes it unpredictable which
will be selected.

To make it predictable, we can break ties using the buildrequestid
instead. This preserves the property of choosing the oldest (as a br
with a lower id will have a submitted_at <= then a br with a higher id).