process: Fix race condition during build cancelling
The race condition happens like this:

 1. buildstep is about to finish. run() is still proceeding
2. build is canceled. BuildStep.interrupt() is invoked. The latter
calls addCompleteLog() which blocks in PlainLog.addContent()
3. run() finishes and proceeds to BuildStep._cleanup_logs(), which
finishes the log that is being written to in (2).
4. PlainLog.addContent() completes and addCompleteLog() calls
PlainLog.finish()
5. PlainLog.finish() upcalls to Log.finish(), which notices that the
log has already been finished in (3) and raises AssertionError.

BuildStep.interrupt() is the only function that may be called outside
the usual lifecycle of a step. BuildStep._cleanup_logs() should wait for
it to complete.