locks: Compare lock owners by id() instead of by instance
Storing lock owners in a list and then using "in" operator to check
existence effectively uses "==" operator for comparisons. This is
incorrect because the code must check not whether two instances are
equal, but whether they are the same object.

To address this and to make code less brittle lock owners are now
tracked by their id.