changes: Implement event recovery using HTTP API in GerritChangeSource
Differently that just combining old GerritChangeSource and
GerritPollingChangeSource, the new implementation restarts SSH client in
case it starts to miss events.

This is important in cases when SSH connection hangs in a way that
simply produces no output. Such cases sometimes happen on intermittent
networks. Note that ServerAliveInterval config option of SSH client does
not help in these particular cases