db: use commit-as-you-go pattern
With SQLA 2.0/future, any connection.execute begin a transaction, no mattern if it's a simple select or not.
Previous preparation to upgrade by scoping write queries in 'with conn.begin():' now causes errors as
a transaction can't be begin when one is already in progress...
But Connection.commit was not available without , thus the rollback now it's activated.