data: fix FieldBase assert when 'values' is a set
The assert is needed as typing with a 'Sequence' also match for 'str' which is not at all a valid type here.
Still typing with a Sequence as maintaining a Union of all Sequence types just to not include str is ... a bit much.
Still assert if str if passed, it would lead to issues elsewhere anyway.